Basics of mesh vertices
MyMesh creates 4 vertices on start, and a MyMeshRenderer draws them in a
single batch via the MyVertex shader.

The MyVertex shader ensures that the quad meshes are always aligned with
the screen and take up a constant percentage of the screen width. This
involves using the meshes original vertex coordinates but using them as
a screen-space offset after determining the origin's clip-space
position.
One thing I didn't know was that you have to multiply the clip-space
offsets by the origin clip's w component. I believe this counter-acts
the perspective divide that happens after the vertex shader whereby the
x, y, and z values are divided by the w component