Add basic ability to click on vertices
This involves raycasting against spheres for each vertex. Scaling the
sphere radii based on distance from camera in view space.

When selections occur, we change a shader property so we can visualize
them differently.

When a click misses everything, the selection is cleared.

Shift click allows for appending to an existing group of selection

Decided on the ISelectionPrimitive to be the basic building block of
selection. Vertex for example iimplements the interface. Our MyMesh
consists of four of such vertices. It is helpful for me to consider the
vertices we are dealing with at the moment as selection primitives as
opposed to geometry primitives, because that is a little different and
involves normals etc. (are verts duplicated for flat shading or not?)