Add axis-constrained translation via hotkey
Basically if you hit G, then X, or G, then Y, then you can move along a
specific axis. This is blender behaviour.

The code is a little funky, in that typically translation happens based
on a delta. i.e. targetPos - curPos = delta. And then we move all of the
selectables by that delta. This is so we don't need to track the start
positions of all the selections. The problem is that when we constrain
movement, we want to take the overall movement from the start pos and
constrain that. So we translate everything back to the start pos with no
constraints, then translate to the target position with constraint. So
every frame we are effectively doing two translations. I'm not super
worried about it atm, but it's kind of funky.