Attempt to clean up hotkey translation code
The previous fix felt weird... Why would we need to set current position
on undo. The reason was the state change method
SetTranslationModeViaHotkey was setting startPos to currentPos and that
was getting called at the beginning and end.
It's useful at the end, but not needed at the beginning, so pulled that
out.

I also tried to rename some things. Instead of translationViaHotkey,
it's now just hotkeyTranlation. The idea is hopefully to make things a
little easier to understand.

For another thing we weren't setting the current position when starting
a translation in this state. The only time we set curPos outside of
active translation is selection changed. This makes me suspect actually
that we will have a similar bug when translating via the gizmo.

Ayayay how the simple act of translation can have so much complexity!