Bugfix: translation with gizmo after undo
As suspected, same problem with translation via gizmo. We weren't
setting the currentPos at the beginning of a gizmo translation because
we don't really have a way to know the beginning of one of those at the
moment. Usually we rely on current position being up to date, because
well it's current!

When we undo, we have to reset the current position since all of our
movements are delta based, and we don't have a good hook to set
currentPos to startPos at the beginning of a translation, so as long as
it stays up to date during translation, selection change, and undo/redo,
we should be ok... That feels like a lot to keep track of. But again, it
works for now. It's just complicated because of the state complexity
with translation.