Fix an issue with box select unexpectedly activating
We have a disableSelection signal which is on while we are doing hotkey
translation, but as soon as that translation is complete, we re-enable
selection. This means the MouseUp and such events may still get fired
even though selection was disabled when the click would have happened.
Hacked in a little hint by setting boxStart to 0 on mouseUp, so if we
get to mouseUp or mouseHeld and we see boxStart is 0, we know the mouse
events should be ignored.

Alternatively we could prevent resetting the disableSelection flag until
we get a mouse up signal, but that feels tedious too. This should be
fine