Copy blender's duplicate+move undo/redo system
In blender, when you duplicate you are automatically moved into a
translate action which we had previously via an event from the
duplication operation with a hook into translate.

But the trick is that when that translation is complete, it doesn't go
in the undo stack as a regular old translation, it actually goes in as a
duplication.

So now in translate we have a special flag for _wasDuplication which
isn't great, but w/e. At the end of every translation we fire a
UnityEvent with the wasDuplication param. This is hooked up to
Duplication's DuplicateTranslationComplete which then adds the
appropriate data to the undo stack. Also note that translate's code only
adds the translate action to the stack if wasDuplication is false.

I hate to special case duplication in translate code, but it is what it
is, and if we find this gets more and more complex, we can deal with it
as it comes