Support incidental deletion of edges
While in vertex selection mode, it is possible that some edges may be
selected, and thus deleted along with the vertices. Deletion code now
handles that case by searching for relevant edges to delete along with
the vertices.

Translation code needed to be updated too because when undoing the
deletion some edges were included, and it didn't know how to position
the gizmo