Selection Manager automatically tracks secondary prims
By secondary primitives I mean primitives that are composed of
lower-level ones. In this case vertices are primary primatives, and
edges are secondary. While selecting vertices the program will
automatically update edges (secondary primitives). This way when you
delete two vertices, the selection will automatically include the edge
between those two.

Note that the code we had previously already handled this for deletion,
but it seemed that that might have been a special case, and so it may be
convenient to just always track secondary and potentially tirtiary
primitives too. We shall see. As I'm typing this out I'm wondering if it
was really necessary.