Ability to click to select edges
Using IQ's capsule intersector function, and setting the radius the same
way we do vertices, but using the center point of the edge to determine
the distance. This might be weird for edges that go a long way
front-to-back in the view frustum. Also just realized scaling of the
vertices might not work as expected in orthographic mode... Nevermind it
appears to just work...