Did the thing that Blender does with edge shading
Namely it will partially shade an edge if you are in vertex mode and
only one vertex is selected, but will fully shade the edge if both are
selected.

Turns out since we are drawing our lines with a vertex-colored shader
you just need to set each vertex based on it's selection, and the gpu
will interpolate the color per-fragment. Woot!