Support box selection for edges
The algorithm is basically, if either point is inside the frustum region
return true, otherwise check if the line segment overlaps the frustum
box in clip space. The box/segment code was ripped from stack overflow.

Also there's a minor bug in the code here that prevents click selection
of a line ssegment that has one vertex behind the camera, and one in
front. Since the vertex behind the camera fails the z-test, the whole
overlapEdges loop will skip to the next edge.