Change how box selection deselects
Since box selection has a live preview, there was some code to deselect
items that were in the previous selection but not in the box currently.
That deselection code got kind of hairy because it depended on the
current selection mode, and it was complicated by the auto-selection of
other primitives that happens.

Now it simply sets the selection to the new selection, and adds the
previous selection if shift is held down. It clears the selection at the
start of each frame. Perhaps not as memory efficient, but this being a
toy and all... Much better to have the code be easier to understand at
this point.