Fix selection when switching from edge to vertex mode
Starting to see some complexities arise in the code when switching
between different modes.

In this case the issue was that vertex mode has this rule where if two
vertices of an edge are selected, that edge is automatically selected.
So added some code in ChangeMode to handle this case.

It appears that ChangeMode is going to need 6 separate cases to consider
for each possible transition.

Also updated the auto-selection on Select/Deselect that adjusts edges
based on vertices. I only really want to do that when we are in vertex
select mode. For one reason without that check we get into a circular
situation, where selecting an edge, selects vertices, which selects
edges. And for the other, it's just not the behaviour we want. If we
select two oposing edges on a quad, we don't want the two adjacent edges
to get auto-selected because their vertices are.

See, complex stuff!