Change how edges are rendered outside of vertex mode
In vertex mode we do this cool edge shading in which edges are colored
based on their vertices selection status. The problem again is if you
have two oposing edges on a quad selected, the two non-selected adjacent
edges will also appear selected by virtue of their vertices being
selected. So if we aren't in vertex selection mode, edge coloring
becomes an all-or-nothing deal based on the selection state of the edge.