Deletion works differently depending on selection mode
Blender lets you perform different kind of deletions in any selection
mode via a drop-down. In our case the deletion mode just depends on the
current selection mode. It's less fully featured but's less work too.

To clarify the reason for this, consider deleting one edge of a quad. In
edge mode, we only want that one edge to be removed. In vertex mode,
where we have two verts of an edge selected, we would expect to delete
not only the verts and the connecting edge, but also other edges that
depend on those verts. This is hinted at with the partial coloring of
edges in vertex selection mode.