Simplify positioning of translation gizmo
Since we auto select vertices when selecting higher order primitives, we
can simply use the vertex positions to place the gizmo. Similar to how
we find the center for Frame Selected, to find the average to position
the transform gizmo, we only need to iterate through the vertices. Also
there will never be a case where only 1 primitive is selected and it's
not a vertex, since the higher order primitives depend on vertices and
assume that they both exist and are selected if a higher order is
selected.