Fix minor issue with translation gizmo placement
After undoing / re-doing a selection operation the translation gizmo
wasn't changing, because we were handling undo/redo by manually
appending / removing items from the selection, which doesn't invoke the
selection changed event.

The fix is to manually invoke the selection change at the end of the
formation undo / redo. It feels a little dirty, but it works.