Update deletion in vertex selection mode
Deletion in vertex selection mode now accounts for faces. Basically if
you delete a vertex, and there is a loop that starts with that vertex,
that implies a polygon uses the vertex, so we need to delete that
polygon.

The trick is that polygons store loop references by indices, so when we
delete a polygon and its loops, we need to shift the existing polygon
loop start indices down to account for it.

More tricky yet is when you undo / redo. The solution here is to rely on
the order in which data exists in the DeleteAction. We are assuming that
there is always a sequence of Poly, Loop, Loop, Loop, Poly, etc...
This means we can always set the loop start index to loops.count on
undo.

This hasn't really been tested at all, so I guess we should try making a
more complex mesh to start with.