Start with a cube mesh
Encapsulated the edge, loop, and poly creation for quad faces of a cube,
and create a basic quad. Not sure if normals are correct. In fact I
think they are wrong, but we can worry about that way later.

This should help test more complex deletion cases, but we shall see.

Also this is kind of cool now that we actually have a 3d mesh