Support deletion of faces via edge selection mode
Similar to vertex selection, we basically go through all the loops and
see if any loops use the edge we are deleting, and if so we delete the
poly by first adding the poly to the delete list, and then the loops.
This uses the same loop index adjustement logic and undo/redo logic as
with the vertex selection mode.