Fix deletion in edge mode, and support face mode
There was a bug in edge mode deletion. Particularly there is code that
checks if deleting an edge should also delete its vertices. Of course if
those vertices are used by another non-deleted edge, then we want to
preserve that vertex. The previously logic wasn't quite right, in that
it would scan all of the remaining edges, however there is the case of a
floating vertex that got skipped. We solve this now by only iterating
over edges that we deleted.

Same pattern in face deletion. We iterate over the deleted loops. First
we check if the edge in a deleted loop is used by another polygon, and
then we check if the vertex in a deleted loop is used by another edge.
If either of those tests fail then we delete the lower order primitive.

Yes, very confusing, and bug prone, but it seems to work question mark.