Fix a bug with loop indices and undo duplication
This fix was pulled from the deletion code. The issue is that whenever
we remove a face (like if we undo the duplication of a face) and we
remove loops, we need to adjust the start index of any polygons whose
loop start is after the deleted one, shifting them down.

The bug would occur when duplication was proceded by deletion, and I'm
not exactly sure why that would trigger it, but the fix seems to work...